+++ src/server/appserver.h
@@ -193,7 +193,7 @@ class AppServer : public SharedTools::QtSingleApplication {
         void sendErrorsCleared(int syncDbId);
         void sendLogUploadStatusUpdated(
             const char state /*state: 'A' for Archiving, 'U' for uploading*/,
-            const int64_t percent);
+            const int percent);
         void sendLogUploadCompleted(bool success, const SyncPath &archivePath);
         void startSyncPals();
         void stopSyncTask(int syncDbId);  // Long task which can block GUI: post-poned in the event loop by means of timer
